% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{chile_plebiscite_df}
\alias{chile_plebiscite_df}
\title{Voting Intentions in the 1988 Chilean Plebiscite}
\format{
A data frame with 2,700 observations and 8 variables:
\describe{
  \item{region}{Region of the respondent (factor with 5 levels: "C", "M", "N", "S", ...)}
  \item{population}{Population size of the respondent's area (integer)}
  \item{sex}{Sex of the respondent (factor: "F", "M")}
  \item{age}{Age in years (integer)}
  \item{education}{Education level (factor with 3 levels: "P", "PS", "S")}
  \item{income}{Income of the respondent (integer)}
  \item{statusquo}{Support for the status quo (numeric scale)}
  \item{vote}{Intended vote (factor with 4 levels)}
}
}
\source{
Data taken from the carData package version 3.0-5
}
\usage{
data(chile_plebiscite_df)
}
\description{
This dataset, chile_plebiscite_df, is a data frame containing information on voting intentions in the 1988 Chilean plebiscite.
The data were collected from a national survey conducted by FLACSO/Chile during April and May of 1988. The dataset consists of
2,700 observations and 8 variables, including demographic details such as region, sex, age, education, and income, as well as
voting preferences and support for the status quo. Some observations contain missing values.
}
\details{
The dataset name has been kept as 'chile_plebiscite_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
ChileDataAPI package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
