% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiledataapi-package.R
\docType{package}
\name{ChileDataAPI}
\alias{ChileDataAPI}
\alias{ChileDataAPI-package}
\title{ChileDataAPI: Access Chilean Data via APIs and Curated Datasets}
\description{
{
This package provides functions to access data from public RESTful APIs including 'FINDIC API', 'REST Countries API', 'World Bank API', and 'Nager.Date', retrieving real-time or historical data related to Chile such as financial indicators, holidays, international demographic and geopolitical indicators, and more. Additionally, the package includes curated datasets related to Chile, covering topics such as human rights violations during the Pinochet regime, electoral data, census samples, health surveys, seismic events, territorial codes, and environmental measurements..
}
}
\details{
ChileDataAPI: Access Chilean Data via APIs and Curated Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access Chilean Data via APIs and Curated Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/chiledataapi}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
