% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiledataapi-package.R
\docType{package}
\name{ChileDataAPI}
\alias{ChileDataAPI}
\alias{ChileDataAPI-package}
\title{ChileDataAPI: Access Chilean Economic, Demographic, Environmental, and Geopolitical Data via RESTful APIs and Curated Datasets}
\description{
{
This package provides functions to access data from the 'FINDIC API' and the 'REST Countries API',related to Chile's financial indicators, international country information, additionally, the package includes curated datasets related to Chile.
}
}
\details{
ChileDataAPI: Access Chilean Economic, Demographic, Environmental, and Geopolitical Data via RESTful APIs and Curated Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access Chilean Economic, Demographic, Environmental, and Geopolitical Data via RESTful APIs and Curated Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/chiledataapi}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
